Flash driver (not used), task and scheduler improvements.

+++ TBuffer.h
@@ -7,18 +7,15 @@
 class TBuffer
 {
 public:
-  TBuffer() :
-    Data(0), Length(0) {}
-  TBuffer(uint8_t* data, uint8_t len) :
-    Data(data), Length(len) {}
+    TBuffer() :
+        Data(0), Length(0) {}
+    TBuffer(uint8_t* data, int len) :
+        Data(data), Length(len) {}
 
-  uint8_t* GetData() const { return Data; }
-  uint8_t GetLength() const { return Length; }
-
-  //start, end
-  //xor, slice
+    uint8_t* GetData() const { return Data; }
+    uint8_t GetLength() const { return Length; }
 
 private:
-  uint8_t* Data;
-  uint8_t Length;
+    uint8_t* Data;
+    int Length;
 };
